release/20200424

承認者： Masataka Fukuba

+++ app/controllers/admin/articles_controller.rb
@@ -19,12 +19,13 @@ module Admin
     end
 
     def edit
-      @article = Article.find(params[:id])
+      @article = current_user.articles.find(params[:id])
     end
 
     def update
-      @article = Article.find(params[:id])
-      if @article.update(article_params)
+      @article = current_user.articles.find(params[:id])
+      @article.attributes = article_params
+      if @article.update_with_history!
         redirect_to articles_path, notice: "記事を更新しました"
       else
         render :edit
@@ -33,6 +34,7 @@ module Admin
 
     def destroy
       @article = Article.find(params[:id])
+      raise ActiveRecord::RecordNotFound  unless @article.user_id == current_user.id
       if @article.destroy
         redirect_to articles_path, notice: "記事を削除しました"
       else
